+++ NimdaConFrontEnd/src/components/Layout/NavBar/index.tsx
@@ -1,0 +1,3 @@
+export { default as BrandSection } from '../../icons/Logo';
+export { default as UserActions } from './UserName';
+export { default } from './NavBar';
