--- NimdaConFrontEnd/src/components/NavBar/index.tsx
@@ -1,3 +1,0 @@
-export { default as BrandSection } from './BrandSection';
-export { default as UserActions } from './UserActions';
-export { default } from '../NavBar';
