+++ NimdaConFrontEnd/src/components/icons/Logo.tsx
