+++ NimdaConFrontEnd/src/components/icons/Logout.tsx
