+++ NimdaConFrontEnd/src/components/icons/Write.tsx
