--- NimdaConFrontEnd/src/domains/User/Login/Page/Layout.style.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-
-export const LayoutContainer = styled.div`
-  display: flex;
-  width: 1440px;
-  padding: 112px 0 320px 0;
-  flex-direction: column;
-  align-items: center;
-  gap: 160px;
-  background: #FAFAF8;
-`;
